requirements.txt
Django==1.8.5
django-filter==0.11.0
djangorestframework==3.2.4
psycopg2==2.6.1
wheel==0.24.0
pdbpp==0.8.3
-e git://git.linaro.org/lava/django-crowd-rest-backend.git@c741f1738786c58253b400ffc27a16967702f84c#egg=django_crowd_rest_backend-dev

